requirements.txt
Django>=5.1,<5.2
django-environ>=0.11
psycopg[binary]>=3.1
mysqlclient>=2.2
gunicorn>=21.2
whitenoise>=6.7
djangorestframework>=3.15
weasyprint>=62.0
pytest>=8.2
pytest-django>=4.8
black>=24.3
ruff>=0.5